requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
PyYAML==6.0
opencensus-ext-azure==1.1.9
opencensus-ext-logging==0.1.1
pyodbc==4.0.35
azure-identity==1.12.0
pandas==2.0.1
sqlalchemy==2.0.12
azure-ai-ml==1.5.0